% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateCpG.R
\name{annotateCpG}
\alias{annotateCpG}
\title{Annotate the CpG ID's with genetic information}
\usage{
annotateCpG(EWAS)
}
\arguments{
\item{EWAS}{An EWAS results data frame with columns for CpG-ID, p-value, standard error and beta coefficient.}
}
\value{
\code{annotateCpG} as an data table.
}
\description{
This funtion annotates the results of an EWAS with the "IlluminaHumanMethylation450kanno.ilmn12.hg19" package.
The function calls the annotation data frame from the IlluminaHumanMethylation450kanno.ilmn12.hg19 bioconductor package
and returns a new data frame, with all the CpG, gene and location information as well as the EWAS model results
}
\details{
IMPORTANT: The CpG identifier column needs to be called "ID"!
}
