% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRfinder.R
\name{DMRfinder}
\alias{DMRfinder}
\title{Differentially methylated regions (DMR) finder}
\usage{
DMRfinder(EWAS, annotate = TRUE, p.column.name = p.column.name,
  beta.column.col = beta.column.col, se.column.col = se.column.col,
  C = 2, lambda = 100, pcutoff = 0.001)
}
\arguments{
\item{EWAS}{An EWAS results data frame with columns for CpG-ID, p-value, standard error and beta coefficient.}

\item{annotate}{The default is that the}

\item{p.column.name}{The name of the p-value column}

\item{beta.column.col}{The name of the beta coefficient column}

\item{se.column.col}{The name of the standard error column}

\item{C}{This is as per \code{DMRcate} package: Scaling factor for bandwidth. Gaussian kernel is calculated where lambda/C = sigma.
Empirical testing shows that, for 450k data when lambda=1000, near- optimal prediction of sequencing-derived DMRs is obtained when C is approxi- mately 2, i.e.
1 standard deviation of Gaussian kernel = 500 base pairs. Should be a lot larger for sequencing data - suggest C=50. Cannot be < 0.2}

\item{lambda}{This is according to the \code{DMRcate} package: Gaussian kernel bandwidth for smoothed-function estimation. Also informs DMR bookend definition;
gaps >= lambda between significant CpG sites will be in separate DMRs. Support is truncated at 5*lambda. Default is 1000 nucleotides. See details for further info.}

\item{pcutoff}{P-value threshold for selecting significan DMRs}
}
\value{
\code{DMRfinder} as a data table.
}
\description{
This function is a wrapper for the \code{DMRcate} bioconductor package to identify differentially methylated regions.
The input needs to be a result file from an EWAS, containing p-values, beta-coefficients and standard errors.
}
\details{
IMPORTANT: The CpG identifier column needs to be called "ID"!
}
