% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSet.R
\name{geneSet}
\alias{geneSet}
\title{Gene set enrichment analysis}
\usage{
geneSet(EWAS, p.val.col = p.val.col, minsize = 100, maxsize = 1000,
  sig.cut = 0.001, plottitle = "title", plot.it = FALSE)
}
\arguments{
\item{EWAS}{An EWAS results data frame with columns for CpG-ID, p-value, standard error and beta coefficient.}

\item{p.val.col}{Columna name for the p-value in the EWAS file}

\item{minsize}{Mimimum amount of genes in GO pathway}

\item{maxsize}{MAximum amount of genes in GO pathway}

\item{sig.cut}{Significance cut off for the gene set enrichment: Only CpGs with a value equal or smaller than this will 
be included}

\item{plottitle}{Title for the heatmap-barplot, if \code{plot.it} is TRUE}

\item{plot.it}{If true, plot will be returned rather than the table with GO terms}
}
\value{
\code{geneSet} as a data table or heatmap-type barplot.
}
\description{
This function is a wrapper for the \code{methylGSA} bioconductor package, which is a very easy to use method
for gene set enrichment analysis
}
\details{
IMPORTANT: The CpG identifier column needs to be called "ID"!
}
