% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manhattan_plot.R
\name{ManhattanPlot}
\alias{ManhattanPlot}
\title{Create a Manhattan plot}
\usage{
ManhattanPlot(EWAS, annotate = TRUE, p.column.name = p.column.name,
  title = title, col.scheme = c("turquoise4", "springgreen4"))
}
\arguments{
\item{EWAS}{An EWAS results data frame with columns for CpG-ID, p-value,}

\item{annotate}{If TRUE (default) this will annotatate the EWAS results file with the gene and position info, using the \code{annotateCpG()} function of
this package}

\item{p.column.name}{This specifies the column for the model p-value in the EWAS result file}

\item{title}{Specify a title for the Manhattan plot}

\item{col.scheme}{Specify two colors for the alternating color scheme of the Manhattan plot}
}
\value{
\code{ManhattanPlot} as an image.
}
\description{
This function utilizes both the \code{annotateCpG()} function of this package, as well as the \code{qqman} CRAN package. It
creates a Manhattan plot with annotated gene names based on the \code{IlluminaHumanMethylation450kanno.ilmn12.hg19} bioconductor package.
}
