% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TKIepiR.R
\docType{package}
\name{TKIepiR-package}
\alias{TKIepiR}
\alias{TKIepiR-package}
\title{TKIepiR: Post-EWAS analysis and visualisation pipeline}
\description{
This package includes all steps that we perform post-EWAS in the
    Diabetes & Obesity group at the Telethon Kids Institute. It contains Manhattan and Volcano plots, 
    DMR with DMRcate, gene annotation and pathway analysis via methylGSA and missMethyl.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/Hobbeist/TKIepiR/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Rauschert \email{Sebastian.Rauschert@telethonkids.org.au}

}
\keyword{internal}
